+++ src/WebJobs.Extensions.OpenAI/Agents/ChatBotPostAttribute.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using System;
 using Microsoft.Azure.WebJobs.Description;
 
 namespace WebJobs.Extensions.OpenAI.Agents;
@@ -20,9 +19,20 @@ public class ChatBotPostAttribute : Attribute
     /// </summary>
     [AutoResolve]
     public string Id { get; }
+
+    /// <summary>
+    /// Gets or sets the OpenAI chat model to use.
+    /// </summary>
+    /// <remarks>
+    /// When using Azure OpenAI, then should be the name of the model <em>deployment</em>.
+    /// </remarks>
+    [AutoResolve]
+    public string? Model { get; set; }
 }
 
 public record ChatBotPostRequest(string UserMessage)
 {
     public string Id { get; set; } = string.Empty;
+
+    public string? Model { get; set; }
 }